test(e2e): replace stale TUI askuser degrade test with disallowed-tools contract
The tui-askuser-degrade e2e test (added in PR #748) still exercised the
Escape-cancel + markdown degrade path through ask-user-degrader.ts, but
that degrader was removed end-to-end in PR #747. At HEAD both handlers
disable AskUserQuestion entirely, so the degrade path is unreachable and
the test asserted a contract that no longer exists.

- Rewrite the test as tui-askuser-disallowed: assert the TUI handler
  spawns claude with `--disallowed-tools AskUserQuestion` as an adjacent
  argv pair, using the fake binary's `start` side-channel event.
- Remove the now-unused FAKE_TUI_EMIT_ASKUSER knob from the fake claude
  TUI binary, the fixture knob mapping, and the askuser:* event kinds
  from the log reader.
- Sync the e2e README and tmux-runtime FINDINGS test matrices.
- docs(server): add missing "request" to the messages.format doc comment
  (matches messageFormatSchema in shared).

Verified: pnpm check, pnpm typecheck, and the full e2e:tui suite (8/8)
all pass.
